app/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
coverage==7.3.2
cryptography==41.0.4
defusedxml==0.8.0rc2
Django==4.2.6
django-filter==23.2
django-phonenumber-field==7.1.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.4
iniconfig==2.0.0
model-bakery==1.15.0
oauthlib==3.2.2
packaging==23.2
phonenumbers==8.13.22
pluggy==1.3.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.8.0
pytest==7.4.2
pytest-cov==4.1.0
pytest-django==4.5.2
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==5.3.1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
urllib3==2.0.6
